sidechains: drop the index-backed wallet
